release 2.28.0-SNAPSHOT

+++ sirma-platform/platform/seip-parent/model-management/definition-core/src/main/java/com/sirma/itt/seip/definition/compile/GenericDefinitionCompilerCallback.java
@@ -44,7 +44,7 @@ public class GenericDefinitionCompilerCallback {
 		definition.setCreationDate(new Date());
 		definition.setLastModifiedDate(new Date());
 		// initialize the revision
-		if (definition.getRevision() == null || definition.getRevision().longValue() == 0) {
+		if (definition.getRevision() == null || definition.getRevision() == 0) {
 			definition.setRevision(1L);
 		}
